requirements.txt
# requirements.txt
dnspython
python-whois
pyopenssl
